Add files via upload
This project is a embedded C project for ECE 36200 (Microprocessor Systems and Interfacing) at Purdue University. 

This alarm clock shows the real time and allows the user to set an alarm. When the alarm rings, a screen will pop up a math question. Until the user gets the question correct, the alarm will keep ringing. For a more awakening effect, the sound changes when the user inputs the wrong question. 

Team members: Seok Hong, Aryan Tyagi, Kubilay Sahin